+++ packages/browser-wallet/src/popup/pages/Settings/Settings.tsx
@@ -32,7 +32,7 @@ export default function Settings() {
     return (
         <div className="settings-page">
             <NavList>
-                <Link className="settings-page__link" to={absoluteRoutes.home.settings.allowList.path}>
+                <Link className="settings-page__link" to={absoluteRoutes.home.settings.allowlist.path}>
                     {t('allowlist')}
                 </Link>
                 <Link className="settings-page__link" to={absoluteRoutes.home.settings.passcode.path}>